+++ biojava3-structure/src/main/java/org/biojava/bio/structure/contact/StructureInterfaceCluster.java
@@ -48,6 +48,6 @@ public class StructureInterfaceCluster implements Serializable {
 		for (StructureInterface interf:members) {
 			area+=interf.getTotalArea();
 		}
-		return area/(double)members.size();
+		return area/members.size();
 	}
 }
